+++ cmd/restic/cmd_rewrite_integration_test.go
